go/mio: introduce mkv.go
