+++ bindings/go/mio/mio.go
@@ -45,7 +45,6 @@ package mio
 // #cgo CFLAGS: -Wno-attributes
 // #cgo LDFLAGS: -L../../../motr/.libs -Wl,-rpath=../../../motr/.libs -lmotr
 // #include <stdlib.h>
-// #include <errno.h> /* EEXIST */
 // #include "lib/types.h"
 // #include "lib/trace.h"
 // #include "motr/client.h"
@@ -105,12 +104,6 @@ type Mio struct {
     objLid  C.ulong
     objPool C.struct_m0_fid
     off     int64
-}
-
-// Mkv provides key-value API to Motr
-type Mkv struct {
-    idxID   C.struct_m0_uint128
-    idx    *C.struct_m0_idx
 }
 
 type slot struct {
@@ -205,15 +198,6 @@ func (mio *Mio) objNew(id string) (err error) {
     return nil
 }
 
-func (mkv *Mkv) idxNew(id string) (err error) {
-    mkv.idxID, err = ScanID(id)
-    if err != nil {
-        return err
-    }
-    mkv.idx = (*C.struct_m0_idx)(C.calloc(1, C.sizeof_struct_m0_idx))
-    return nil
-}
-
 // GetPool returns the pool the object is located at.
 func (mio *Mio) GetPool() string {
     if mio.obj == nil {
@@ -252,131 +236,6 @@ func (mio *Mio) open(sz uint64) error {
     mio.off = 0
 
     return nil
-}
-
-// Open opens Mkv index for key-value operations.
-func (mkv *Mkv) Open(id string, create bool) error {
-    if mkv.idx != nil {
-        return errors.New("index is already opened")
-    }
-
-    err := mkv.idxNew(id)
-    if err != nil {
-        return err
-    }
-
-    C.m0_idx_init(mkv.idx, &C.container.co_realm, &mkv.idxID)
-
-    if create { // Make sure it's created
-        var op *C.struct_m0_op
-        rc := C.m0_entity_create(nil, &mkv.idx.in_entity, &op)
-        if rc != 0 {
-            mkv.Close()
-            return fmt.Errorf("failed to set create op: %d", rc)
-        }
-        C.m0_op_launch(&op, 1)
-        rc = C.m0_op_wait(op, bits(C.M0_OS_FAILED,
-                                   C.M0_OS_STABLE), C.M0_TIME_NEVER)
-        if rc == 0 {
-            rc = C.m0_rc(op)
-        }
-        C.m0_op_fini(op)
-        C.m0_op_free(op)
-
-        if rc != 0 && rc != -C.EEXIST {
-            return fmt.Errorf("index create failed: %d", rc)
-        }
-    }
-
-    return nil
-}
-
-// Close closes Mkv index releasing all the resources
-// that were allocated for it.
-func (mkv *Mkv) Close() error {
-    if mkv.idx == nil {
-        return errors.New("index is not opened")
-    }
-    C.m0_idx_fini(mkv.idx)
-    C.free(unsafe.Pointer(mkv.idx))
-    mkv.idx = nil
-
-    return nil
-}
-
-func (mkv *Mkv) doIdxOp(name uint32, key []byte, value []byte,
-                        update bool) ([]byte, error) {
-    if mkv.idx == nil {
-        return nil, errors.New("index is not opened")
-    }
-
-    var k, v C.struct_m0_bufvec
-    if C.m0_bufvec_empty_alloc(&k, 1) != 0 {
-        return nil, errors.New("failed to allocate key bufvec")
-    }
-    defer C.m0_bufvec_free2(&k)
-    if C.m0_bufvec_empty_alloc(&v, 1) != 0 {
-        return nil, errors.New("failed to allocate value bufvec")
-    }
-    // we should cleanup the buffer allocated by GET op
-    if name == C.M0_IC_GET {
-        defer C.m0_bufvec_free(&v)
-    } else {
-        defer C.m0_bufvec_free2(&v)
-    }
-
-    *k.ov_buf = unsafe.Pointer(&key[0])
-    *k.ov_vec.v_count = C.ulong(len(key))
-    if name != C.M0_IC_GET {
-        *v.ov_buf = unsafe.Pointer(&value[0])
-        *v.ov_vec.v_count = C.ulong(len(value))
-    }
-
-    var op  *C.struct_m0_op
-    var rcI  C.int32_t
-    flags := C.uint(0)
-    if name == C.M0_IC_PUT && update {
-        flags = C.M0_OIF_OVERWRITE
-    }
-    rc := C.m0_idx_op(mkv.idx, name, &k, &v, &rcI, flags, &op)
-    if rc != 0 {
-        return nil, fmt.Errorf("failed to init index op: %d", rc)
-    }
-
-    C.m0_op_launch(&op, 1)
-    rc = C.m0_op_wait(op, bits(C.M0_OS_FAILED,
-                               C.M0_OS_STABLE), C.M0_TIME_NEVER)
-    if rc == 0 {
-        rc = C.m0_rc(op)
-    }
-    C.m0_op_fini(op)
-    C.m0_op_free(op)
-
-    if rc != 0 {
-        return nil, fmt.Errorf("op failed: %d", rc)
-    }
-    if rcI != 0 {
-        return nil, fmt.Errorf("index op failed: %d", rcI)
-    }
-
-    if name == C.M0_IC_GET {
-        value = make([]byte, *v.ov_vec.v_count)
-        copy(value, pointer2slice(*v.ov_buf, int(*v.ov_vec.v_count)))
-    }
-
-    return value, nil
-}
-
-// Put puts key-value into the index.
-func (mkv *Mkv) Put(key []byte, value []byte, update bool) error {
-    _, err := mkv.doIdxOp(C.M0_IC_PUT, key, value, update)
-    return err
-}
-
-// Get gets value from the index by key.
-func (mkv *Mkv) Get(key []byte) ([]byte, error) {
-    value, err := mkv.doIdxOp(C.M0_IC_GET, key, nil, false)
-    return value, err
 }
 
 // Open opens Mio object for reading ant/or writing. The size
